[jit] Use computed goto to dispatch to next instruction in interpreter.
Summary:
Refactor the interpreter loop a bit to use `goto` to jump to the next
instruction at the end of each instruction body, only on supported compiler.
In some microbenchmark this can reduce the branch misprediction rate of
interpreter loop by 10x. Altough there isn't much improvement for running
time in real cases, it's potentially still beneficial to do this optimization
and refactor the loop to be more readable in the meanwhile.

Test Plan:
CI

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: fcf34ec3d87940784228aad3f4261634e48d87ed
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60211