Enable jit tracing to parametrization and add jit tests
This PR fixes the tracing in the parametrizations.
The current resolution is that when tracing is performed while caching is enabled, we throw an error.
Without caching, the tracing should work properly (tests added).

Currently, the parametrizations don't support scripting.
This PR introduces the same logic as with the tracing (throw error if caching).
However, the scripting itself cannot enabled due to the use of the generator expressions in the parametrizations.
Added TODO to fix it.

Test Plan:

```python
python test/test_jit.py -- TestParametrization
```

ghstack-source-id: fe4aad590f7f92ee5dd1a06edb0b659b01e5cd94
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60969